+++ AFD0001400375031063605426574000140REP_C.txt
@@ -1,4 +1,83 @@
+00000000162024-04-25T16:40:00-030002
+00000000262024-04-25T16:43:00-030002
+00000000362024-08-02T09:55:00-030002
+00000000422024-08-02T10:45:00-030083384316215   10542657400014000083384316215JUSTICA FEDERAL NO AMAPA                                                                                                                              RODOVIA NORTE SUL SN                                                                                93f1
+00000000542024-08-02T10:50:00-03002024-08-02T10:50:00-030083384316215c352
+00000000662024-08-02T10:56:00-030002
+00000000722024-08-02T12:07:00-030083384316215   10542657400014000083384316215JUSTICA FEDERAL DE PRIMEIRO GRAU NO AMAPA                                                                                                             RODOVIA NORTE SUL SN                                                                                ba32
+00000000852024-08-02T12:08:00-0300I083384316215Bruno Alexandre Soares Prestes                      000S83384316215c34c
+00000000952024-08-02T12:08:00-0300A083384316215Bruno Alexandre Soares Prestes                      000S833843162150e1d
+00000001052024-08-02T12:09:00-0300A083384316215Bruno Alexandre Soares Prestes                      100S833843162158da5
+00000001152024-08-02T12:09:00-0300A083384316215Bruno Alexandre Soares Prestes                      100S83384316215eb31
+00000001262024-08-02T12:11:00-030005
+00000001332024-08-02T12:17:00-030008338431621540fc
+00000001432024-08-02T12:19:00-03000833843162150cae
+00000001522024-08-02T12:51:00-030083384316215   10542657400014000000000000000JUSTICA FEDERAL DE PRIMEIRO GRAU NO AMAPA                                                                                                             RODOVIA NORTE SUL S/N                                                                               65db
+00000001622024-08-02T12:51:00-030083384316215   10542657400014000000000000000JUSTICA FEDERAL DE PRIMEIRO GRAU NO AMAPA                                                                                                             RODOVIA NORTE SUL S/N                                                                               bd39
+00000001752024-08-02T13:24:00-0300I812312312345USUARIO TESTE                                       0000833843162151bd3
+00000001832024-08-02T13:28:00-03000833843162155667
+00000001952024-08-02T13:37:00-0300E812312312345USUARIO TESTE                                       0000833843162155c4c
+00000002052024-08-02T13:37:00-0300I081231231234USUARIO TESTE                                       000083384316215986a
+00000002152024-08-02T14:34:00-0300E081231231234USUARIO TESTE                                       0000833843162156e17
+00000002262024-08-02T14:36:00-030002
+00000002352024-08-02T14:40:00-0300A083384316215Bruno Prestes                                       100S83384316215331a
+00000002462024-08-05T10:37:00-030002
+00000002532024-08-05T16:12:00-03000833843162154179
+00000002662024-08-05T21:14:00-030002
+00000002762024-08-05T22:37:00-030002
+00000002862024-08-06T10:42:00-030002
+00000002932024-08-06T10:49:00-030008338431621557f1
+00000003032024-08-06T13:57:00-03000833843162158d4a
+00000003132024-08-06T14:24:00-0300083384316215b05b
+00000003232024-08-06T14:32:00-03000833843162155d27
+00000003332024-08-06T16:39:00-0300083384316215aba2
+00000003462024-08-06T18:11:00-030002
+00000003562024-08-06T22:30:00-030002
+00000003662024-08-06T23:25:00-030002
+00000003732024-08-07T10:48:00-0300083384316215e0ad
+00000003832024-08-07T12:54:00-0300083384316215a3e7
+00000003962024-08-11T16:45:00-030002
+00000004032024-08-16T12:20:00-0300083384316215262f
+00000004132024-08-16T12:21:00-0300083384316215bb8b
+00000004232024-08-16T12:27:00-0300083384316215ffaa
+00000004352024-08-16T12:38:00-0300I000630482217GABRIEL FELIPE                                      100083384316215df67
+00000004432024-08-16T12:38:00-0300000630482217e98b
+00000004562024-08-16T21:31:00-030002
+00000004632024-08-16T21:35:00-0300083384316215626c
+00000004732024-08-16T21:36:00-03000833843162159110
+00000004832024-08-18T16:20:00-03000833843162155af8
+00000004962024-09-23T13:13:00-030002
+00000005032024-09-23T13:15:00-03000833843162154716
+00000005132024-09-23T14:07:00-0300083384316215a296
+00000005232024-09-23T14:10:00-03000833843162157886
+00000005362024-09-23T14:11:00-030002
+00000005432024-09-23T14:11:00-0300083384316215a969
+00000005562024-09-23T14:11:00-030006
+00000005662024-09-23T14:14:00-030002
+00000005742024-09-23T14:18:00-03002024-09-23T14:18:00-030083384316215ae58
+00000005862024-10-02T10:36:00-030002
+00000005932024-10-02T11:24:00-0300083384316215cdf2
+00000006032024-10-02T11:25:00-0300083384316215f171
+00000006132024-10-02T11:29:00-03000833843162153578
+00000006232024-10-02T11:40:00-0300083384316215a9d1
+00000006332024-10-02T12:01:00-03000833843162156cef
+00000006432024-10-02T12:09:00-030008338431621593d5
+00000006532024-10-02T12:09:00-0300083384316215391d
+00000006632024-10-02T17:52:00-030008338431621506df
+00000006762024-10-03T06:34:00-030002
+00000006832024-10-03T10:07:00-0300083384316215b21b
+00000006932024-10-03T10:52:00-0300083384316215ee04
+00000007032024-10-03T12:06:00-03000833843162150704
+00000007152024-10-03T12:56:00-0300A083384316215Bruno Prestes                                       100S83384316215257c
+00000007232024-10-04T11:04:00-03000833843162156e78
 00000007362024-10-08T10:48:00-030002
 00000007432024-10-08T10:51:00-0300083384316215e63b
-
+00000007562024-10-21T10:11:00-030002
+00000007662025-03-06T11:42:00-030002
+00000007732025-03-06T11:44:00-03000833843162158a36
+00000007832025-03-06T12:25:00-03000833843162158d35
+00000007952025-03-06T12:32:00-0300A000630482217GABRIEL FELIPE                                      10008338431621542c5
+9999999990000000040000000370000000020000000120000000240000000009
+oDlMlyDU+bWPJ0lrjk8FG6qgruh4QrqZnJjHDaQtUzTgkfvu0G24f7xJAZLAVzlvWBbYOasQbBYsMYFey2I5gw==            
+000000000110542657400014000000000000000JUSTICA FEDERAL DE PRIMEIRO GRAU NO AMAPA                                                                                                             000140037503106362024-04-252025-03-062025-03-11T11:43:00-0300003108238299000129iDClass Bio Prox              2ced
 AFD0001400375031063605426574000140REP_C.txt
