+++ nsr.json
@@ -1,1 +1,1 @@
-{"nsr":"000000074","tipoRegistro":"3 - MARCAÇÃO DE PONTO","dataHora":"2024-10-08T10:51:00-0300","cpf":"83384316215","crc16":"1DED"}
+{"nsr":"000000013","tipoRegistro":"3 - MARCAÇÃO DE PONTO","dataHora":"2024-08-02T12:17:00-0300","cpf":"83384316215","crc16":"078E"}
